requirements.txt
paho-mqtt==2.1.0
protobuf==4.25.4
tahutils==0.1.9
